[Starts #163807952] Restructure Tests
- Add model tests
- Restructure App
- Break down views to own files
