+++ api/v1/models/user_model.py
@@ -1,0 +1,2 @@
+class UserModel:
+    pass
